Extract email bounce webhook controller
